+++ app/controllers/question_flow_controller.rb
@@ -38,26 +38,6 @@ protected
     @step ||= Flow::Handler.step_from_url_fragment(params[:step_url_fragment])
   end
 
-  def tag_from(step)
-    return if Flow::Handler::STEPS.fetch(step)[:tag].nil?
-
-    Flow::Handler::STEPS.fetch(step)[:tag]
-  end
-
-  def last_tag_in_group?(tag)
-    remaining_steps = Steps::Helper.remaining_steps_for(session_data, step)
-
-    return if remaining_steps.blank?
-
-    remaining_tags = []
-    remaining_steps.map do |remaining|
-      remaining_tags << tag_from(remaining)
-    end
-    return unless !remaining_tags.compact.include?(tag) && tag_from(step) == tag
-
-    true
-  end
-
   def change_answers_loop?
     params[:controller] == "change_answers"
   end
