+++ app/controllers/strategy/change_answers_strategy.rb
@@ -1,0 +1,4 @@
+module Strategy
+  class ChangeAnswersStrategy
+  end
+end
