--- app/controllers/strategy/check_answers_strategy.rb
@@ -1,4 +1,0 @@
-module Strategy
-  class CheckAnswersStrategy
-  end
-end
