+++ app/controllers/strategy/form_handler_strategy.rb
@@ -1,4 +1,7 @@
 module Strategy
   class FormHandlerStrategy
+    def next_step session_data, step
+      Steps::Helper.next_step_for(session_data, step)
+    end
   end
 end
